% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_signatures.R
\name{cell_signatures}
\alias{cell_signatures}
\title{Cell_Signatures}
\usage{
cell_signatures(
  coords,
  expression,
  signatures,
  file.path.prefix,
  cexsize = 1,
  type = "UMAP",
  subset = 30:45
)
}
\arguments{
\item{coords}{Dataframe containing 2 columns, each for X and Y coordinates with rows as cells}

\item{expression}{Dataframe of gene expression values, rows as genes and cells as columns}

\item{signatures}{The signatures file, with each column as a seperate list of genes (signature) with headers}

\item{file.path.prefix}{File path, including sample name, for plot to be saved as with the appropriate suffix. Example: "/Users/Documents/BPDCN123"}

\item{cexsize}{The value that determines "cex" in the plot. Generated as part of the QC and Filter workflow but can be supplied independently. Default is 1}

\item{type}{The type of plots to generate/the type of coordinates given. "UMAP" (default) and "tSNE" accepted}

\item{subset}{Range of numeric values indicating which 16 plots to generate as the seperate "selected" plots. Default is 30:45 for the Griffin signatures}
}
\description{
This function allows you to provide dimensionality coordinates and a gene by cell matrix to get cell signature overlays.
}
\examples{
cell_signatures()
}
\keyword{UMAP}
\keyword{cell}
\keyword{plot}
\keyword{signature}
\keyword{tSNE}
