% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_and_filter.R
\name{qc_and_filter}
\alias{qc_and_filter}
\title{QC_and_Filter}
\usage{
qc_and_filter(
  directories,
  genome = "GRCh38",
  suffix = "default",
  min.umis = 500,
  min.genes = 250,
  max.mito = 0.3,
  sig = T,
  dense = F
)
}
\arguments{
\item{directories}{List of directories containing CellRanger output for QC and filtering. Example single directory path is "/Volumes/broad_vangalenlab/depasquale/CellRanger/210628_GeneCounts"}

\item{genome}{Desired genome: "GRCh38" and "GRCm38" are accepted, default is "GRCh38"}

\item{suffix}{Suffix for directory where the output of this function will be stored and where the input sample sheet is located. Default is the "QC" suffix.}

\item{min.umis}{QC filter for minimum number of Unique Molecular Identifiers (UMIs) detected per cell. Default is 500}

\item{min.genes}{QC filter for minimum number of genes detected per cell. Default is 250}

\item{max.mito}{QC filter to remove cells that exceed this fraction of mitochondrial alignments. Default is 0.3}

\item{sig}{Boolean for whether or not to plot cell type signatures. Default is T}

\item{dense}{Boolean for whether or not to save dense matrices alongside sparse matrices. Default is F}
}
\description{
This function allows you to provide a list of paths leading to the QC sample sheet (Erica's pipeline Step 4) to perform quality control and filtering.
}
\examples{
qc_and_filter()
}
\keyword{filter}
\keyword{qc}
