% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_transfer_data.R
\name{seurat_transfer_data}
\alias{seurat_transfer_data}
\title{Seurat Transfer Data}
\usage{
seurat_transfer_data(
  reference,
  datasets,
  results_name = "results",
  ordering_vector = "none",
  cluster_names = "none",
  cluster_colors = "none"
)
}
\arguments{
\item{reference}{Path to a Seurat object that is an integrated reference}

\item{datasets}{List of directories containing a Seurat object Example single directory path is "~/Documents/Projects/Data_files_temp/Donor/SBM1014/Seurat.rds"}

\item{results_name}{The name you would prefer to call the output directory. Default is "results"}

\item{ordering_vector}{A numeric vector containing the order you would like your clusters to appear in the plot, starting with 1. Default is "none", which retains the original order.}

\item{cluster_names}{A character vector containing the names of the clusters, in the original order. Default is "none" or cluster numbers to be used.}

\item{cluster_colors}{A character vector containing hex codes for the colors of cluster, in the original order. Default is "none" for random colors.}
}
\description{
This function allows you to classify cells in the list of datasets using the reference with the Seurat transferData() function
}
\examples{
seurat_transfer_data()
}
\keyword{classification}
\keyword{seurat}
\keyword{transfer}
